Factorize structutil/serializable in a sane way
