+++ src/rootfilespec/bootstrap/strings.py
@@ -1,6 +1,6 @@
+from rootfilespec.buffer import ReadBuffer
 from rootfilespec.dispatch import DICTIONARY
-from rootfilespec.serializable import serializable
-from rootfilespec.structutil import Members, ReadBuffer, ROOTSerializable
+from rootfilespec.serializable import Members, ROOTSerializable, serializable
 
 
 @serializable
