+++ src/rootfilespec/rntuple/RFrame.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Generic, TypeVar
 
-from rootfilespec.structutil import Members, ReadBuffer, ROOTSerializable
+from rootfilespec.buffer import ReadBuffer
+from rootfilespec.serializable import Members, ROOTSerializable
 
 Item = TypeVar("Item", bound=ROOTSerializable)
