+++ src/rootfilespec/rntuple/pagelist.py
@@ -1,5 +1,6 @@
 from typing import Annotated
 
+from rootfilespec.buffer import DataFetcher, ReadBuffer
 from rootfilespec.rntuple.envelope import (
     ENVELOPE_TYPE_MAP,
     REnvelope,
@@ -7,13 +8,8 @@ from rootfilespec.rntuple.envelope import (
 from rootfilespec.rntuple.pagelocations import ClusterLocations
 from rootfilespec.rntuple.RFrame import ListFrame, RecordFrame
 from rootfilespec.rntuple.RPage import RPage
-from rootfilespec.serializable import serializable
-from rootfilespec.structutil import (
-    DataFetcher,
-    Fmt,
-    Members,
-    ReadBuffer,
-)
+from rootfilespec.serializable import Members, serializable
+from rootfilespec.structutil import Fmt
 
 
 @serializable
